usage_logging/requirements.txt
google-cloud-bigquery
boto3
azure-identity
azure-monitor-query
pandas